CI: Setup repo for CI
#### Problem

The associated-token-account program, interface, and tests are in the
repo, but they aren't properly exercised through CI.

#### Summary of changes

Get the whole repo setup! Contains the following changes:

* move `program-test` into `program/tests` for simplicity
* remove unneeded `use_latest_*` functions in program-test
* add CI for build / lint / format / test the program and interface
* add rust-specific scripts and re-use them
* update maintainers to Anza

NOTE: The interface crate is still called `client`, but since it
functions as more of an interface crate, I moved it into `interface`.
The crate rename can happen later.