harness: Rename to spl-associated-token-account-mollusk-harness (#159)
#### Problem

In order to publish the new ATA crate, we also need to publish the
mollusk harness.

#### Summary of changes

Make the mollusk harness crate name conform a bit more to the rest of
the repo, by calling it `spl-associated-token-account-mollusk-harness`.
It's a bit of a mouthful, but it's certainly clear!

Also, rename the directory from test_harness -> mollusk_harness to go
with the crate name.